Correction in ProductionNameValidationRuleTest + New Rule: ProductDiscountValidationRule
