+++ src/main/java/com/javaguru/shoppinglist/service/validation/rules/ProductPriceValidationRule.java
@@ -1,0 +1,17 @@
+package com.javaguru.shoppinglist.service.validation.rules;
+
+import com.javaguru.shoppinglist.domain.Product;
+import com.javaguru.shoppinglist.service.validation.ProductValidationException;
+import com.javaguru.shoppinglist.service.validation.ProductValidationRule;
+
+import java.math.BigDecimal;
+
+public class ProductPriceValidationRule implements ProductValidationRule {
+    @Override
+    public void validate(Product product) {
+        checkNotNull(product);
+        if (product.getPrice().compareTo(BigDecimal.ZERO) < 0) {
+            throw new ProductValidationException("Price should be less than 0");
+        }
+    }
+}
