+++ src/test/java/com/javaguru/shoppinglist/service/validation/ProductNameValidationRuleTest.java
@@ -1,6 +1,7 @@
 package com.javaguru.shoppinglist.service.validation;
 
 import com.javaguru.shoppinglist.domain.Product;
+import com.javaguru.shoppinglist.service.validation.rules.ProductNameValidationRule;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -8,14 +9,12 @@ import static org.junit.Assert.*;
 public class ProductNameValidationRuleTest {
 
     @Test
-    public void test1(){
+    public void testNameLengthValidation(){
         ProductNameValidationRule r = new ProductNameValidationRule();
         Product product = new Product();
         product.setName("ap");
         try {
-            System.out.println("before");
             r.validate(product);
-
             fail();
         } catch (Exception e) {
             assertEquals(e.getMessage(), "Name should be not less, than 3 letters and not more, than 100");
@@ -23,7 +22,7 @@ public class ProductNameValidationRuleTest {
 
     }
     @Test
-    public void test2() {
+    public void testEmptyNameValidation() {
         ProductNameValidationRule r = new ProductNameValidationRule();
         Product product = new Product();
         product.setName("");
@@ -31,7 +30,7 @@ public class ProductNameValidationRuleTest {
             r.validate(product);
 //            fail();
         } catch (Exception e) {
-            assertNull(e.getMessage(), "Product name must be not null");
+            assertEquals(e.getMessage(), "Product name must be not empty");
         }
     }
